# Conflicts:
#	contexts/AppContext.js
#	hooks/useStats.js

+++ services/sessionService.js
@@ -11,24 +11,26 @@ export const refreshSessions = async (uid) => {
     const { sessions, fullRoundSessions } = await getUserSessionsByID(uid);
     const newData = await fetchUserData(uid);
     if (!newData.sessionsUpdated) {
-        const files = await RNFS.readDir(sessionDirectory);
-        for (const file of files) {
-            const content = await RNFS.readFile(file.path, 'utf8');
-            const sessionData = JSON.parse(content);
-            const sessionId = file.name.split('.')[0];
-            if (!sessionData.synced) {
-                await setDoc(doc(firestore, `users/${uid}/sessions`, sessionId), { ...sessionData, synced: true });
-                await RNFS.unlink(file.path);
+        if (await RNFS.exists(sessionDirectory)) {
+            const files = await RNFS.readDir(sessionDirectory);
+            for (const file of files) {
+                const content = await RNFS.readFile(file.path, 'utf8');
+                const sessionData = JSON.parse(content);
+                const sessionId = file.name.split('.')[0];
+                if (!sessionData.synced) {
+                    await setDoc(doc(firestore, `users/${uid}/sessions`, sessionId), { ...sessionData, synced: true });
+                    await RNFS.unlink(file.path);
+                }
             }
-        }
-        const fullRoundFiles = await RNFS.readDir(fullRoundDirectory);
-        for (const file of fullRoundFiles) {
-            const content = await RNFS.readFile(file.path, 'utf8');
-            const fullRoundData = JSON.parse(content);
-            const sessionId = file.name.split('.')[0];
-            if (!fullRoundData.synced) {
-                await setDoc(doc(firestore, `users/${uid}/fullRoundSessions`, sessionId), { ...fullRoundData, synced: true });
-                await RNFS.unlink(file.path);
+            const fullRoundFiles = await RNFS.readDir(fullRoundDirectory);
+            for (const file of fullRoundFiles) {
+                const content = await RNFS.readFile(file.path, 'utf8');
+                const fullRoundData = JSON.parse(content);
+                const sessionId = file.name.split('.')[0];
+                if (!fullRoundData.synced) {
+                    await setDoc(doc(firestore, `users/${uid}/fullRoundSessions`, sessionId), { ...fullRoundData, synced: true });
+                    await RNFS.unlink(file.path);
+                }
             }
         }
         await updateUserData(uid, { sessionsUpdated: true });
